requirements.txt
google-generativeai==0.4.0
streamlit==1.31.1  # or any newer version that fits
streamlit-option-menu==0.3.12
pip install youtube-transcript-api
pip install pathlib
